Include response canceled by Chrome or user action
