+++ index.js
@@ -17,6 +17,7 @@ const {
   toNameValuePairs
 } = require('./lib/util');
 const populateEntryFromResponse = require('./lib/entryFromResponse');
+const finalizeEntry = require('./lib/finalizeEntry');
 
 const defaultOptions = {
   includeResourcesFromDiskCache: false,
@@ -374,45 +375,7 @@ module.exports = {
               continue;
             }
 
-            const timings = entry.timings || {};
-            timings.receive = formatMillis(
-              (params.timestamp - entry._requestTime) * 1000 -
-                entry.__receiveHeadersEnd
-            );
-            entry.time =
-              max(0, timings.blocked) +
-              max(0, timings.dns) +
-              max(0, timings.connect) +
-              max(0, timings.send) +
-              max(0, timings.wait) +
-              max(0, timings.receive);
-
-            // For cached entries, Network.loadingFinished can have an earlier
-            // timestamp than Network.dataReceived
-
-            // encodedDataLength will be -1 sometimes
-            if (params.encodedDataLength >= 0) {
-              const response = entry.response;
-              if (response) {
-                response._transferSize = params.encodedDataLength;
-                response.bodySize = params.encodedDataLength;
-
-                if (
-                  isHttp1x(response.httpVersion) &&
-                  response.headersSize > -1
-                ) {
-                  response.bodySize -= response.headersSize;
-                }
-
-                const compression = Math.max(
-                  0,
-                  response.content.size - response.bodySize
-                );
-                if (compression > 0) {
-                  response.content.compression = compression;
-                }
-              }
-            }
+            finalizeEntry(entry, params)
           }
           break;
 
@@ -483,6 +446,16 @@ module.exports = {
               );
               continue;
             }
+
+            if (params.errorText === "net::ERR_ABORTED") {
+              finalizeEntry(entry, params)
+              debug(
+                `Loading was canceled due to Chrome or a user action for requestId ${
+                  params.requestId
+                }.`
+              );
+              continue;
+            };
 
             // This could be due to incorrect domain name etc. Sad, but unfortunately not something that a HAR file can
             // represent.
